requirements.txt
# Core ML Libraries
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
scipy==1.12.0

# NLP Libraries
nltk==3.8.1
spacy==3.7.2

# Deep Learning (Python 3.12 compatible)
tensorflow==2.18.0

# MLOps Tools
mlflow==2.10.2
dvc==3.48.0

# API & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-multipart==0.0.9

# Visualization
matplotlib==3.8.2
seaborn==0.13.2

# Utilities
joblib==1.3.2
python-dotenv==1.0.1
pyyaml==6.0.1

# Development & Testing
pytest==8.0.0
black==24.1.1
flake8==7.0.0
nltk
spacy